Fixed permissions separation by roles
Now permissions of all roles are merged before authorization. Perviously
they all was checked one-by-one. This behaviout lead to the error when
for example:
User has one role which allow Read and another which allow Update, but
action requires both Read and Update permissions at once, so user won't
be authorized for desired action.
In this case technically user has both this permissions, but they are in
different roles and that was not handled properly before.

Also fixed a bug when roles in resources with permissions == 0
considered as unset. (which actually means that users with this roles
can't do anything with this resource)